refactor: moved participa components and utils
