+++ src/app/ui/index.ts
@@ -2,6 +2,7 @@ export * from "./breadcrumbs"
 export * from "./button"
 export * from "./cards"
 export * from "./form"
+export * from "./menu/Menu"
 export * from "./Pagination"
 export * from "./proyectoByInvestigador/ProyectoViewerByInvestigador"
 export * from "./Search"
