--- src/app/utils/participa/add.ts
@@ -1,8 +1,0 @@
-import { addParticipa } from "./crudActions"
-import { ParticipaCommand } from "./participaCommand"
-
-export class AddParticipaCommand extends ParticipaCommand {
-  execute() {
-    addParticipa(this.participa)
-  }
-}
