--- src/app/utils/participa/index.ts
@@ -1,3 +1,0 @@
-export * from "./add"
-export * from "./participaCommand"
-export * from "./remove"
