--- src/app/utils/participa/remove.ts
@@ -1,8 +1,0 @@
-import { removeParticipa } from "./crudActions"
-import { ParticipaCommand } from "./participaCommand"
-
-export class DeleteParticipaCommand extends ParticipaCommand {
-  execute() {
-    removeParticipa(this.participa)
-  }
-}
