+++ src/participa/command/index.ts
@@ -1,0 +1,3 @@
+export * from "./addCommand"
+export * from "./participaCommand"
+export * from "./removeCommand"
