+++ src/participa/menu/index.ts
@@ -1,0 +1,1 @@
+export * from "./AddParticipante"
