+++ src/proyectos/form/create/useProyectoCreate.ts
@@ -5,8 +5,7 @@ import type { FormEvent } from "react"
 import { useState } from "react"
 import { useDebouncedCallback } from "use-debounce"
 
-import type { ParticipaType } from "@/app/utils"
-import { AddParticipaCommand } from "@/app/utils"
+import { AddParticipaCommand, type ParticipaType } from "@/participa"
 import type { ProyectoType } from "@/proyectos/types"
 import type { ProyectoValidationErrors } from "@/proyectos/utils"
 import { addProyecto, getProyectoErrors } from "@/proyectos/utils"
